+++ code/Dockerfile.gpu
@@ -1,3 +1,4 @@
+# FIXME: outdated Ubuntu version.
 FROM ubuntu:18.04
 
 RUN apt update && apt install -y python3
