+++ code/requirements.tests.txt
@@ -1,0 +1,4 @@
+-r requirements.txt
+mock
+pytest
+pytest-cov
